carrental/requirements.txt
Pillow==9.5.0
asgiref==3.6.0
Django==4.1.4
djangorestframework==3.14.0
sqlparse==0.4.3
python-decouple==3.8


